chore(tooling): separate diagnostics executables
