+++ .github/workflows/docs-pages.yml
@@ -8,7 +8,6 @@ on:
       - '.github/workflows/**'
       - '.nvmrc'
       - 'config/check-release-profile.mjs'
-      - 'config/check-diagnostic-catalog.mjs'
       - 'config/diagnostics/**'
       - 'config/docs/**'
       - 'config/release-profile.json'
@@ -17,6 +16,7 @@ on:
       - 'upgradeGuide.md'
       - 'package-lock.json'
       - 'package.json'
+      - 'scripts/diagnostics/**'
   pull_request:
   workflow_dispatch:
