+++ package.json
@@ -80,7 +80,7 @@
   "scripts": {
     "build": "rollup -c rollup.config.mjs --noConflict",
     "check:browser-profile": "node config/check-browser-profile.mjs",
-    "check:diagnostics": "node config/check-diagnostic-catalog.mjs",
+    "check:diagnostics": "node scripts/diagnostics/check-catalog.mjs",
     "check:dist": "npm run build && node config/check-dist.mjs",
     "check:release-promotion": "node config/release/preflight.mjs",
     "check:release-profile": "node config/check-release-profile.mjs",
