+++ scripts/diagnostics/catalog.mjs
